% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/interact_rc.R
\name{interact_rc}
\alias{interact_rc}
\title{Run Interactive Rogers-Castro App}
\usage{
interact_rc()
}
\value{
No return value, called for interactive widget
}
\description{
Run an interactive Rogers-Castro app. Use interactive sliders to see how parameters affect the Rogers-Castro age schedules.
}
\examples{
\dontrun{
interact_rc()
}
}
